+++ packages/video_player/video_player_platform_interface/lib/video_player_platform_interface.dart
@@ -78,6 +78,17 @@ abstract class VideoPlayerPlatform extends PlatformInterface {
     throw UnimplementedError('setVolume() has not been implemented.');
   }
 
+  /// Returns true if caching is supported for the mimetype of the network video url.
+  Future<bool> isCacheSupportedForNetworkMedia(String url) {
+    throw UnimplementedError(
+        'isCacheSupportedForNetworkMedia() has not been implemented.');
+  }
+
+  /// Clears the cached videos.
+  Future<void> clearCache(int textureId) {
+    throw UnimplementedError('clearCache() has not been implemented.');
+  }
+
   /// Sets the video position to a [Duration] from the start.
   Future<void> seekTo(int textureId, Duration position) {
     throw UnimplementedError('seekTo() has not been implemented.');
@@ -102,6 +113,11 @@ abstract class VideoPlayerPlatform extends PlatformInterface {
   Future<void> setMixWithOthers(bool mixWithOthers) {
     throw UnimplementedError('setMixWithOthers() has not been implemented.');
   }
+
+  /// Sets additional options on web
+  Future<void> setWebOptions(int textureId, VideoPlayerWebOptions options) {
+    throw UnimplementedError('setWebOptions() has not been implemented.');
+  }
 }
 
 class _PlaceholderImplementation extends VideoPlayerPlatform {}
@@ -122,14 +138,16 @@ class DataSource {
   ///
   /// The [package] argument must be non-null when the asset comes from a
   /// package and null otherwise.
-  DataSource({
-    required this.sourceType,
-    this.uri,
-    this.formatHint,
-    this.asset,
-    this.package,
-    this.httpHeaders = const <String, String>{},
-  });
+  DataSource(
+      {required this.sourceType,
+      this.uri,
+      this.formatHint,
+      this.asset,
+      this.package,
+      this.httpHeaders = const <String, String>{},
+      this.maxCacheSize,
+      this.maxFileSize,
+      this.enableCache = false});
 
   /// The way in which the video was originally loaded.
   ///
@@ -142,6 +160,24 @@ class DataSource {
   /// This will be in different formats depending on the [DataSourceType] of
   /// the original video.
   final String? uri;
+
+  /// The URI to the video file.
+  ///
+  /// This will be in different formats depending on the [DataSourceType] of
+  /// the original video.
+  final int? maxCacheSize;
+
+  /// The URI to the video file.
+  ///
+  /// This will be in different formats depending on the [DataSourceType] of
+  /// the original video.
+  final int? maxFileSize;
+
+  /// Enable cache for media.
+  ///
+  /// This will be in different formats depending on the [DataSourceType] of
+  /// the original video.
+  final bool? enableCache;
 
   /// **Android only**. Will override the platform's generic file format
   /// detection with whatever is set here.
@@ -365,7 +401,7 @@ class DurationRange {
 /// [VideoPlayerOptions] can be optionally used to set additional player settings
 @immutable
 class VideoPlayerOptions {
-  /// set additional optional player settings
+  /// Set additional optional player settings
   // TODO(stuartmorgan): Temporarily suppress warnings about not using const
   // in all of the other video player packages, fix this, and then update
   // the other packages to use const.
@@ -373,6 +409,7 @@ class VideoPlayerOptions {
   VideoPlayerOptions({
     this.mixWithOthers = false,
     this.allowBackgroundPlayback = false,
+    this.webOptions,
   });
 
   /// Set this to true to keep playing video in background, when app goes in background.
@@ -385,4 +422,86 @@ class VideoPlayerOptions {
   /// Note: This option will be silently ignored in the web platform (there is
   /// currently no way to implement this feature in this platform).
   final bool mixWithOthers;
+
+  /// Additional web controls
+  final VideoPlayerWebOptions? webOptions;
+}
+
+/// [VideoPlayerWebOptions] can be optionally used to set additional web settings
+@immutable
+class VideoPlayerWebOptions {
+  /// [VideoPlayerWebOptions] can be optionally used to set additional web settings
+  const VideoPlayerWebOptions({
+    this.controls = const VideoPlayerWebOptionsControls.disabled(),
+    this.allowContextMenu = true,
+    this.allowRemotePlayback = true,
+  });
+
+  /// Additional settings for how control options are displayed
+  final VideoPlayerWebOptionsControls controls;
+
+  /// Whether context menu (right click) is allowed
+  final bool allowContextMenu;
+
+  /// Whether remote playback is allowed
+  final bool allowRemotePlayback;
+}
+
+/// [VideoPlayerWebOptions] can be used to set how control options are displayed
+@immutable
+class VideoPlayerWebOptionsControls {
+  /// Enables controls and sets how the options are displayed
+  const VideoPlayerWebOptionsControls.enabled({
+    this.allowDownload = true,
+    this.allowFullscreen = true,
+    this.allowPlaybackRate = true,
+    this.allowPictureInPicture = true,
+  }) : enabled = true;
+
+  /// Disables control options. Default behavior.
+  const VideoPlayerWebOptionsControls.disabled()
+      : enabled = false,
+        allowDownload = false,
+        allowFullscreen = false,
+        allowPlaybackRate = false,
+        allowPictureInPicture = false;
+
+  /// Whether native controls are enabled
+  final bool enabled;
+
+  /// Whether downloaded control is displayed
+  ///
+  /// Only applicable when [controlsEnabled] is true
+  final bool allowDownload;
+
+  /// Whether fullscreen control is enabled
+  ///
+  /// Only applicable when [controlsEnabled] is true
+  final bool allowFullscreen;
+
+  /// Whether playback rate control is displayed
+  ///
+  /// Only applicable when [controlsEnabled] is true
+  final bool allowPlaybackRate;
+
+  /// Whether picture in picture control is displayed
+  ///
+  /// Only applicable when [controlsEnabled] is true
+  final bool allowPictureInPicture;
+
+  /// A string representation of disallowed controls
+  String get controlsList {
+    final List<String> controlsList = <String>[];
+    if (!allowDownload) {
+      controlsList.add('nodownload');
+    }
+    if (!allowFullscreen) {
+      controlsList.add('nofullscreen');
+    }
+    if (!allowPlaybackRate) {
+      controlsList.add('noplaybackrate');
+    }
+
+    return controlsList.join(' ');
+  }
 }