+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,6 +1,11 @@
-## NEXT
+## 6.2.0
 
+* Adds caching functionality for network video.
 * Updates minimum supported SDK version to Flutter 3.3/Dart 2.18.
+* Adds web options to customize control's list and displaying context menu.
+* Adds caching functionality for network video.
+* Updates minimum supported SDK version to Flutter 3.3/Dart 2.18.
+* Adds caching functionality for network video.
 
 ## 6.1.0